move sims with depinject to tests
