+++ tests/integration/distribution/appconfig.go
@@ -1,10 +1,12 @@
-package testutil
+package distribution_test
 
 import (
 	_ "cosmossdk.io/x/auth"           // import as blank for app wiring
 	_ "cosmossdk.io/x/auth/tx/config" // import as blank for app wiring
 	_ "cosmossdk.io/x/bank"           // import as blank for app wiring
+	_ "cosmossdk.io/x/distribution"   // import as blank for app wiring
 	_ "cosmossdk.io/x/mint"           // import as blank for app wiring
+	_ "cosmossdk.io/x/protocolpool"   // import as blank for app wiring
 	_ "cosmossdk.io/x/staking"        // import as blank for app wiring
 
 	"github.com/cosmos/cosmos-sdk/testutil/configurator"
@@ -19,5 +21,7 @@ var AppConfig = configurator.NewAppConfig(
 	configurator.TxModule(),
 	configurator.ConsensusModule(),
 	configurator.GenutilModule(),
+	configurator.DistributionModule(),
 	configurator.MintModule(),
+	configurator.ProtocolPoolModule(),
 )
