+++ tests/integration/evidence/app_config.go
@@ -1,12 +1,11 @@
-package testutil
+package evidence_test
 
 import (
 	_ "cosmossdk.io/x/auth"           // import as blank for app wiring
 	_ "cosmossdk.io/x/auth/tx/config" // import as blank for app wiring
 	_ "cosmossdk.io/x/bank"           // import as blank for app wiring
-	_ "cosmossdk.io/x/distribution"   // import as blank for app wiring
-	_ "cosmossdk.io/x/mint"           // import as blank for app wiring
-	_ "cosmossdk.io/x/protocolpool"   // import as blank for app wiring
+	_ "cosmossdk.io/x/evidence"       // import as blank for app wiring
+	_ "cosmossdk.io/x/slashing"       // import as blank for app wiring
 	_ "cosmossdk.io/x/staking"        // import as blank for app wiring
 
 	"github.com/cosmos/cosmos-sdk/testutil/configurator"
@@ -18,10 +17,9 @@ var AppConfig = configurator.NewAppConfig(
 	configurator.AuthModule(),
 	configurator.BankModule(),
 	configurator.StakingModule(),
+	configurator.SlashingModule(),
 	configurator.TxModule(),
 	configurator.ConsensusModule(),
+	configurator.EvidenceModule(),
 	configurator.GenutilModule(),
-	configurator.DistributionModule(),
-	configurator.MintModule(),
-	configurator.ProtocolPoolModule(),
 )
