+++ tests/integration/mint/app_config.go
@@ -1,13 +1,10 @@
-package testutil
+package mint
 
 import (
 	_ "cosmossdk.io/x/auth"           // import as blank for app wiring
 	_ "cosmossdk.io/x/auth/tx/config" // import as blank for app wiring
 	_ "cosmossdk.io/x/bank"           // import as blank for app wiring
-	_ "cosmossdk.io/x/distribution"   // import as blank for app wiring
 	_ "cosmossdk.io/x/mint"           // import as blank for app wiring
-	_ "cosmossdk.io/x/protocolpool"   // import as blank for app wiring
-	_ "cosmossdk.io/x/slashing"       // import as blank for app wiring
 	_ "cosmossdk.io/x/staking"        // import as blank for app wiring
 
 	"github.com/cosmos/cosmos-sdk/testutil/configurator"
@@ -19,11 +16,8 @@ var AppConfig = configurator.NewAppConfig(
 	configurator.AuthModule(),
 	configurator.BankModule(),
 	configurator.StakingModule(),
-	configurator.SlashingModule(),
 	configurator.TxModule(),
 	configurator.ConsensusModule(),
 	configurator.GenutilModule(),
 	configurator.MintModule(),
-	configurator.DistributionModule(),
-	configurator.ProtocolPoolModule(),
 )
