+++ tests/integration/staking/module_test.go
@@ -1,4 +1,4 @@
-package mint_test
+package staking
 
 import (
 	"testing"
@@ -7,25 +7,26 @@ import (
 
 	"cosmossdk.io/depinject"
 	"cosmossdk.io/log"
-	authkeeper "cosmossdk.io/x/auth/keeper"
+	authKeeper "cosmossdk.io/x/auth/keeper"
 	authtypes "cosmossdk.io/x/auth/types"
-	"cosmossdk.io/x/mint/testutil"
-	"cosmossdk.io/x/mint/types"
+	"cosmossdk.io/x/staking/types"
 
 	simtestutil "github.com/cosmos/cosmos-sdk/testutil/sims"
 )
 
 func TestItCreatesModuleAccountOnInitBlock(t *testing.T) {
-	var accountKeeper authkeeper.AccountKeeper
-
+	var accountKeeper authKeeper.AccountKeeper
 	app, err := simtestutil.SetupAtGenesis(
 		depinject.Configs(
-			testutil.AppConfig,
+			AppConfig,
 			depinject.Supply(log.NewNopLogger()),
 		), &accountKeeper)
 	require.NoError(t, err)
 
 	ctx := app.BaseApp.NewContext(false)
-	acc := accountKeeper.GetAccount(ctx, authtypes.NewModuleAddress(types.ModuleName))
+	acc := accountKeeper.GetAccount(ctx, authtypes.NewModuleAddress(types.BondedPoolName))
+	require.NotNil(t, acc)
+
+	acc = accountKeeper.GetAccount(ctx, authtypes.NewModuleAddress(types.NotBondedPoolName))
 	require.NotNil(t, acc)
 }
