--- x/staking/app_test.go
@@ -1,129 +1,0 @@
-package staking_test
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/require"
-
-	"cosmossdk.io/collections"
-	"cosmossdk.io/core/header"
-	"cosmossdk.io/depinject"
-	"cosmossdk.io/log"
-	"cosmossdk.io/math"
-	authtypes "cosmossdk.io/x/auth/types"
-	bankKeeper "cosmossdk.io/x/bank/keeper"
-	stakingKeeper "cosmossdk.io/x/staking/keeper"
-	"cosmossdk.io/x/staking/testutil"
-	"cosmossdk.io/x/staking/types"
-
-	"github.com/cosmos/cosmos-sdk/crypto/keys/ed25519"
-	"github.com/cosmos/cosmos-sdk/crypto/keys/secp256k1"
-	simtestutil "github.com/cosmos/cosmos-sdk/testutil/sims"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	moduletestutil "github.com/cosmos/cosmos-sdk/types/module/testutil"
-)
-
-var (
-	priv1 = secp256k1.GenPrivKey()
-	addr1 = sdk.AccAddress(priv1.PubKey().Address())
-	priv2 = secp256k1.GenPrivKey()
-	addr2 = sdk.AccAddress(priv2.PubKey().Address())
-
-	valKey          = ed25519.GenPrivKey()
-	commissionRates = types.NewCommissionRates(math.LegacyZeroDec(), math.LegacyZeroDec(), math.LegacyZeroDec())
-)
-
-func TestStakingMsgs(t *testing.T) {
-	genTokens := sdk.TokensFromConsensusPower(42, sdk.DefaultPowerReduction)
-	bondTokens := sdk.TokensFromConsensusPower(10, sdk.DefaultPowerReduction)
-	genCoin := sdk.NewCoin(sdk.DefaultBondDenom, genTokens)
-	bondCoin := sdk.NewCoin(sdk.DefaultBondDenom, bondTokens)
-
-	acc1 := &authtypes.BaseAccount{Address: addr1.String()}
-	acc2 := &authtypes.BaseAccount{Address: addr2.String()}
-	accs := []simtestutil.GenesisAccount{
-		{GenesisAccount: acc1, Coins: sdk.Coins{genCoin}},
-		{GenesisAccount: acc2, Coins: sdk.Coins{genCoin}},
-	}
-
-	var (
-		bankKeeper    bankKeeper.Keeper
-		stakingKeeper *stakingKeeper.Keeper
-	)
-
-	startupCfg := simtestutil.DefaultStartUpConfig()
-	startupCfg.GenesisAccounts = accs
-
-	app, err := simtestutil.SetupWithConfiguration(
-		depinject.Configs(
-			testutil.AppConfig,
-			depinject.Supply(log.NewNopLogger()),
-		),
-		startupCfg, &bankKeeper, &stakingKeeper)
-	require.NoError(t, err)
-	ctxCheck := app.BaseApp.NewContext(true)
-
-	require.True(t, sdk.Coins{genCoin}.Equal(bankKeeper.GetAllBalances(ctxCheck, addr1)))
-	require.True(t, sdk.Coins{genCoin}.Equal(bankKeeper.GetAllBalances(ctxCheck, addr2)))
-
-	// create validator
-	description := types.NewDescription("foo_moniker", "", "", "", "")
-	createValidatorMsg, err := types.NewMsgCreateValidator(
-		sdk.ValAddress(addr1).String(), valKey.PubKey(), bondCoin, description, commissionRates, math.OneInt(),
-	)
-	require.NoError(t, err)
-
-	headerInfo := header.Info{Height: app.LastBlockHeight() + 1}
-	txConfig := moduletestutil.MakeTestTxConfig()
-	_, _, err = simtestutil.SignCheckDeliver(t, txConfig, app.BaseApp, headerInfo, []sdk.Msg{createValidatorMsg}, "", []uint64{0}, []uint64{0}, true, true, priv1)
-	require.NoError(t, err)
-	require.True(t, sdk.Coins{genCoin.Sub(bondCoin)}.Equal(bankKeeper.GetAllBalances(ctxCheck, addr1)))
-
-	ctxCheck = app.BaseApp.NewContext(true)
-	validator, err := stakingKeeper.GetValidator(ctxCheck, sdk.ValAddress(addr1))
-	require.NoError(t, err)
-
-	require.Equal(t, sdk.ValAddress(addr1).String(), validator.OperatorAddress)
-	require.Equal(t, types.Bonded, validator.Status)
-	require.True(math.IntEq(t, bondTokens, validator.BondedTokens()))
-
-	// edit the validator
-	description = types.NewDescription("bar_moniker", "", "", "", "")
-	editValidatorMsg := types.NewMsgEditValidator(sdk.ValAddress(addr1).String(), description, nil, nil)
-
-	headerInfo = header.Info{Height: app.LastBlockHeight() + 1}
-	_, _, err = simtestutil.SignCheckDeliver(t, txConfig, app.BaseApp, headerInfo, []sdk.Msg{editValidatorMsg}, "", []uint64{0}, []uint64{1}, true, true, priv1)
-	require.NoError(t, err)
-
-	ctxCheck = app.BaseApp.NewContext(true)
-	validator, err = stakingKeeper.GetValidator(ctxCheck, sdk.ValAddress(addr1))
-	require.NoError(t, err)
-	require.Equal(t, description, validator.Description)
-
-	// delegate
-	require.True(t, sdk.Coins{genCoin}.Equal(bankKeeper.GetAllBalances(ctxCheck, addr2)))
-	delegateMsg := types.NewMsgDelegate(addr2.String(), sdk.ValAddress(addr1).String(), bondCoin)
-
-	headerInfo = header.Info{Height: app.LastBlockHeight() + 1}
-	_, _, err = simtestutil.SignCheckDeliver(t, txConfig, app.BaseApp, headerInfo, []sdk.Msg{delegateMsg}, "", []uint64{1}, []uint64{0}, true, true, priv2)
-	require.NoError(t, err)
-
-	ctxCheck = app.BaseApp.NewContext(true)
-	require.True(t, sdk.Coins{genCoin.Sub(bondCoin)}.Equal(bankKeeper.GetAllBalances(ctxCheck, addr2)))
-	_, err = stakingKeeper.Delegations.Get(ctxCheck, collections.Join(addr2, sdk.ValAddress(addr1)))
-	require.NoError(t, err)
-
-	// begin unbonding
-	beginUnbondingMsg := types.NewMsgUndelegate(addr2.String(), sdk.ValAddress(addr1).String(), bondCoin)
-	headerInfo = header.Info{Height: app.LastBlockHeight() + 1}
-	_, _, err = simtestutil.SignCheckDeliver(t, txConfig, app.BaseApp, headerInfo, []sdk.Msg{beginUnbondingMsg}, "", []uint64{1}, []uint64{1}, true, true, priv2)
-	require.NoError(t, err)
-
-	// delegation should exist anymore
-	ctxCheck = app.BaseApp.NewContext(true)
-	_, err = stakingKeeper.Delegations.Get(ctxCheck, collections.Join(addr2, sdk.ValAddress(addr1)))
-	require.ErrorIs(t, err, collections.ErrNotFound)
-
-	// balance should be the same because bonding not yet complete
-	require.True(t, sdk.Coins{genCoin.Sub(bondCoin)}.Equal(bankKeeper.GetAllBalances(ctxCheck, addr2)))
-}
